clean up super-duped parseDateAttrs tests
